changed audio
I didnt like the way the audio was working when declared as a global variable. It would skip the next question if the sound effect was still overlapping. Se I declared each audio variable within the several button functions, and I like the way the sounds work much better now.